refactor: extract shared token-propagation report pipeline
CC002/CC003/CC004 ended in a verbatim ~35-line tail (token-argument
check, scope walk, expression-tree guard, overload check, diagnostic
construction). That tail is now
CancellationTokenHelpers.ReportIfTokenNotPropagated; each analyzer is
rule-specific gating plus one call. No behavior change; all 200 tests
pin it.
